Profile & Registration Fixed Template CreateTask, Detail Task, EditTask
