+++ Tubes_Progin_1/public_html/Dashboard.html
@@ -84,7 +84,7 @@ and open the template in the editor.
                 </form>
                 <br/><br/>
                 <a onclick="showEdit();" class="button">edit</a><br/>
-            </div>
+		</div>
             
             <datalist id="assignee">
                 <option value="Frilla" />
