+++ Tubes_Progin_2/css.css
@@ -250,10 +250,11 @@ input.liusername{
     position:absolute;
     background-color:transparent;
     border:0px;
-    left: 700px;
+    left: 680px;
     padding: 50px;
     margin-left: 100px;
     overflow: hidden;
+	visibility: visible;
     background-image: url('img/kertas.png');
     width: 350px;
     height: 443px;
@@ -297,7 +298,7 @@ textArea {
     margin-left: 100px;
     width: 350px;
     height: 443px;
-    visibility: hidden;
+    visibility: visible;
 }
 
 #overlay {
