+++ Tubes_Progin_2/register.php
@@ -7,8 +7,8 @@
 	$dob = $_POST['regdate'];
 	$password = md5($_POST['regpassword1']);
 	$email = $_POST['regemail'];
+	$target = "img/".$_FILES["regfile"]["name"];
 	move_uploaded_file($_FILES["regfile"]["tmp_name"],"img/".$_FILES["regfile"]["name"]);
-	$target = "img/"."EndyDoankavatar.docx"; //session username sbg "target"avatar.docx
 	$insert_sql = "INSERT INTO `user`(`username`, `fullname`, `avatar`, `birthday`, `email`, `password`) VALUES ('$username','$fullname', '$target', '$dob', '$email', '$password')";
 	mysqli_query($con,$insert_sql);
 	$_SESSION['username'] = $username;
