+++ Tubes_Progin_2/rincitask.php
@@ -40,16 +40,24 @@ and open the template in the editor.
 		<div id="wanted">
 			<img src="img/kertas2.png">
 		</div>
-		<div class="tugas" id="buattugas"><br/>
-			<form id="createForm" method="post" action="create.php" enctype="multipart/form-data">
-                Name: <div class="nama"><input type="text" id="namaTask"></div><br/>
-                Attachment: <div class="attachment"><input type="file"></div><br/>
-                Deadline: <div class="deadline"><input type="date"></div><br/>
-                Assignee: <div class="asignee"><input type="text"></div><br/>
-                Tag: <div class="tag"> <input type="text"></div> <br/>
-                <br/>
-				<input type="submit" id="regbutton" name="submit" value="Register" disabled="disabled">
-			</form>
-		</div>
+		<div class="tugas" id="rincitugas">
+                Name: Nama Tugas <br/>
+                Attachment: 
+                    <div class="attachment">
+                        <a href="img/file.zip">file.zip</a><br/>
+                        <a href="img/badge.png">picture.png</a><br/>
+                    </div><br/>
+                Deadline: 17-12-2014<br/>
+                Assignee: <a href="" class="asignee">Timo</a>, <a href="" class="asignee">Stefan</a>, <a href="" class="asignee">Frilla</a><br/>
+                Tag: <a href="" class="tag">dangerous</a>, <a href="" class="tag">novice</a> <br/>
+                <br/>Comment:<br/>
+                <div class="komentar">Dangerous criminal. Proceed with caution.</div><br/>
+                <form>
+                    <textArea></textarea>
+                    <input type="button" name="submit" value="submit">
+                </form>
+                <br/><br/>
+                <a onclick="showEdit();" class="button">edit</a><br/>
+            </div>
     </body>
 </html>
